backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
PyJWT==2.6.0
bcrypt==4.0.1
motor==3.7.1

# Test iÃ§in
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0
python-jose==3.3.0
psutil==5.9.5

# Adblock listesi (requests)
requests==2.31.0
